polling package refactored
