+++ src/main/java/com/configcat/ConfigService.java
@@ -6,6 +6,9 @@ import com.configcat.log.ConfigCatLogMessages;
 import com.configcat.log.ConfigCatLogger;
 import com.configcat.models.Entry;
 import com.configcat.models.Setting;
+import com.configcat.polling.AutoPollingMode;
+import com.configcat.polling.LazyLoadingMode;
+import com.configcat.polling.PollingMode;
 import org.apache.commons.codec.binary.Hex;
 import org.apache.commons.codec.digest.DigestUtils;
