+++ src/main/java/com/configcat/polling/ManualPollingMode.java
@@ -1,0 +1,8 @@
+package com.configcat.polling;
+
+public class ManualPollingMode implements PollingMode {
+    @Override
+    public String getPollingIdentifier() {
+        return "m";
+    }
+}
