+++ src/test/java/com/configcat/AutoPollingPolicyTest.java
@@ -7,6 +7,8 @@ import com.configcat.cache.SingleValueCache;
 import com.configcat.hooks.ConfigCatHooks;
 import com.configcat.log.ConfigCatLogger;
 import com.configcat.log.LogLevel;
+import com.configcat.polling.PollingMode;
+import com.configcat.polling.PollingModes;
 import okhttp3.OkHttpClient;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
