+++ src/test/java/com/configcat/ConfigFetcherTest.java
@@ -7,6 +7,7 @@ import com.configcat.log.ConfigCatLogger;
 import com.configcat.log.LogLevel;
 import com.configcat.models.Config;
 import com.configcat.models.Entry;
+import com.configcat.polling.PollingModes;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import okhttp3.OkHttpClient;
