+++ src/test/java/com/configcat/ManualPollingPolicyTest.java
@@ -4,6 +4,8 @@ import com.configcat.cache.InMemoryCache;
 import com.configcat.cache.NullConfigCache;
 import com.configcat.hooks.ConfigCatHooks;
 import com.configcat.log.ConfigCatLogger;
+import com.configcat.polling.PollingMode;
+import com.configcat.polling.PollingModes;
 import okhttp3.OkHttpClient;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
