+++ src/test/java/com/configcat/VariationIdTests.java
@@ -1,5 +1,6 @@
 package com.configcat;
 
+import com.configcat.polling.PollingModes;
 import okhttp3.OkHttpClient;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
